Big bootstrap re-write.
This implements the BCa method for bootstrapping the confidence intercal of a statistic.
It borrows heavily from scipy.stats.bootstrap but makes it possible to distribute some of the tasks,
namely calculating the boostrap distribution and the jackknife resampling.

The BCa method is supposedly better at dealing with skewed/biased distributions.

Since calculating the CI of distribution fits is problematic (wrong) this new method skips this and instead
calculates the probabiloty ratio for each resampled distribution. These prob. ratios are then what makes up the
bootstrap distribution.

This is considerably slower compared to the old method, the jackknife introduces n new distribution fits
(n being the length of the sample) which is fairly slow. But this works for now.