Update notebooks and structure
Moves noteboos into a separate folder. Notebooks are also updated to the current state of the complete analysis.
This entails a rewrite of how the bootstrap is done. We no longer apply a bootstrap to the distribution fit but instead do for the probability ratio "directly".
I.e. we now compute the probability ratio n number of times instead of performing the distribution fit n times and then selecting a CI of the fit paramaters.