Initial preprocessing functions.
This adds the modules preprocessing and config. Config contains a small helper function to read a configuration file (.yml) located in the root folder of any attribution project. This contains path, names, variables and such so that the analysis can be customised without changing code in the attribution package.

preprocessing current holds three function which prepares iris cubes for GridClim, EOBS and the CORDEX ensemble. The preparation consist of selecting a common timspan and area of interest which is specified either in the config.yml or by passing arguments to the function. The cubes are saved in a folder specified in config.yml or by an argument. Naming of the files works in the same way.

It is likely that more prepare_* functions will be added as more datasets are explored.