compute_index: Add dimensional padding, lazify, remove index_catalog
Some changes to compute_index. Since the index_catalog cant be re-used and is fairly expensive to initialize,  the prepared index is instead passed to the function.

Climix also require data to be lazy, so cube is lazified before climix.

Finally, currently the consecutive indices require the cube to be 3d. So, non 3d cubes are now "padded" with dummy dimensions to fulfill this.