Probability ratio CI for daily scaling
This adds a new bootstrap function to calculate the probability ratio CI
using shifted daily data: `prob_raio_ds_ci`. It computes a distribution
of probability ratios from n random realisations of the original cube
and then uses scipy to bootstrap the confidence interval of the median
of said distribution.

It comes with some utilities:

A new windowed resampling function `daily_resampling_windows`, making it possible to resample each while still conserving
seasonality and trends. This work by defining a window around each day,
with a size in both days and years, e.g. 15 of July 2010 can be randomly
resampled from 25th of June to 5th of August 2007 - 2013.

`prepare_resampled_cubes`: Pre-processing an "original" cube, and a shifted cube from the resampled data. Used along with an multiprocessing pool to process cubes in parallel.

`calc_prob_ratio_ds`: Computing the probabilty ratio of an event with
predetermined threshold based on a current and a counterfactual cube.