Quantile shifting data
This adds the possibility to shift daily data according to quantile
regression. It extends the concept of monthly median shifting by
computing a regression coefficient for 30 quantiles in each month,
resulting in 360 regression coefficients. Any day is then linearly shifted
according to its month and which quantile it belongs to.

It adds two new functions responsible for the quantile regression and
the shifting. These can then used through `prepare_resampled_cubes`
and `calc_prob_ratio_ds_ci` with the `quantile_shift` kwarg.