prob_ratio_ci revamp
Some redesigning of the prob_ratio_ci (formerly bootstrap_mp). This is
now purely used for bootstrapping the confidence interval of the
probability ratio. Bigger changes:

* ensemble kwarg: Turns of the resampling and assumes the (2d) input data
represent an ensemble, and uses these members for the uncertainty.
* A new helper function _boot_helper, which is used internally to first
evaluate which distribution suits the data and then compute the
probability ratio.
* Removal of the use of the Bca method. We no instead simply calculate
the quantiles for the theta_hat_b distribution.