Fall back to use a single distribution and ensemble kwargs.
Changing the distribution between resampled datasets seems to cause some
issues.
Reverting to using a single distribution for all resamples for now.

With the `ensemble` kwarg in `calc_prob_ratio_ds_ci`, the random
resampling is turned off and the input data, now assumed to be an
ensemble, is used for the uncertainty.