Fixing the resampling and adding a regression cutoff.
There was a bug in the previous resampling method, resulting in a not
true re-shuffle. This should now be fixed. The resampling is also now
done in windows, so a random year will come from neighboring years.

Also added the possibility for a cutoff point for when regression
coefficents should be treated as 0 (when the regression p-value exceeds
a threshold).