Big daily scaling rewrite
Major rewrite of the daily scaling, resulting in a substantial
speedup (~35 min. vs. ~50 s.). The biggest difference is that we no
longer handle 1000s of cubes but instead a single cube with an added
dimensions holding the:

- random realisations

- If data is shifted. (Not shifted, shifted (all regressions), shifted
  (significant regressions))

So a cube with shape [m] that is resampled 50 times will have shape [50, 3, m].
By doing this expensive calls, such as cube.extract is minimised. By holding all
realisations in the same array makes it possible to vectorise operations on
that were previously done sequentially. Some operations that handle single
realisations are mapped using the processor pool and "parallelised". This also decreases Climix calls
from n_resamples to 1.