De-trending data and confidence interval on regression coefficients
This makes it possible to compute the prob. ratio CI using de-trended
data. This is enabled by the `detrend` kwarg to `prob_ratio_ci`. The
de-trending is done using the linear regression to GMST. One can think
of this as moving all years to the climate of the last year of the
dataset. The shifting/scaling for the prob. ratio is unchanged.

Alos added confidence interval of the coefficients to the outputs of `compute_cube_regression`.